pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obdiag"
version = "5.0.0"
description = "OceanBase Diagnostic Tool - cluster diagnosis, log gathering and analysis"
readme = "LICENSE"
license = { text = "Mulan PSL v2" }
requires-python = ">=3.11"
authors = [
    { name = "OceanBase" }
]
keywords = ["oceanbase", "diagnostic", "obdiag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "certifi>=2024.7.4",
    "cffi>=2.0.0",
    "charset-normalizer>=2.0.12",
    "cryptography>=46.0.3",
    "idna>=3.7",
    "Jinja2>=3.1.6",
    "MarkupSafe>=2.0.1",
    "paramiko>=3.4.1",
    "pyasn1>=0.4.8",
    "pycparser>=2.21",
    "PyNaCl>=1.6.2",
    "requests>=2.32.4",
    "rsa>=4.8",
    "six>=1.16.0",
    # Data processing
    "tabulate>=0.8.9",
    "PyYAML>=6.0.2",
    "prettytable>=3.5.0",
    "xmltodict>=0.13.0",
    "sqlparse>=0.5.0",
    # Database & Cloud
    "docker>=6.1.3",
    "PyMySQL>=1.1.1",
    "kubernetes>=30.1.0",
    # CLI & UI
    "pwinput>=1.0.3",
    "colorama>=0.4.6",
    "progressbar>=2.5",
    "halo>=0.0.31",
    # Tools
    "inspect2>=0.1.2",
    "sqlgpt-parser>=0.0.1a5",
    "pyminizip>=0.2.6",
    "plotly>=5.24.1",
    # obdiag agent — tui layer is vendored in src/handler/agent/tui/
    # Core agent runtime (replaces deepagents-cli as a single dependency)
    "deepagents>=0.5.0",
    "deepagents-acp>=0.0.5",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-mcp-adapters>=0.0.1",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=0.0.1",
    "langgraph-sdk>=0.1.0",
    "langsmith>=0.1.0",
    "aiosqlite>=0.19.0",
    "textual>=0.80.0",
    "textual-autocomplete>=3.0.0",
    "textual-speedups>=0.1.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "python-dotenv>=1.0.0",
    "packaging>=23.0",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "markdown-it-py>=3.0.0",
    "markdownify>=0.11.0",
    "pyperclip>=1.8.0",
    "pygments>=2.15.0",
    "uuid-utils>=0.6.0",
    "tavily-python>=0.3.0",
]

[project.optional-dependencies]
# Build binary (PyInstaller)
build = [
    "pyinstaller>=4.3",
]
# macOS: use pyzipper instead of pyminizip (pyminizip has compilation issues on macOS)
macos = [
    "pyzipper>=0.3.6",
]

[tool.setuptools.package-dir]
"src" = "src"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src" = ["conf/*.yml", "conf/*.yaml"]

[project.scripts]
obdiag = "src.main:main"
